Add "jumpref" support to the CouchDBViewPager.
A "jumpref" is a type of pageref that does not affect the view results
nor the user-actioned paging but allows you to jump into the middle of a
view.

A classic use case is an A-Z index to access a list of names, although
a jumpkey can be anything that CouchDB can handle.