Change signature of CouchDB pagers to allow for customisation args.
In the future, I want to add a simple optimisation to say that the key
is unique in which case the docid does not need to be added to the
pageref.

Unfortunately, the __init__ signature was __init__(..., **view_args)
making it very hard to add keyword args without clashing with legitimate
view args. So, I've changed the sig to be __init__(..., view_args=None)
instead. That breaks existing code but better to do it now rather than
when there's more stuff using it.