Serialize the pageref to and from a string.
The (undocumented) idea of a pageref was for it to be an opaque value
only understood by the pager implementation. The ListPager assumed it
was an int already, meaning something else outside the ListPager had to
know to turn it into an int and therefore what serialization the
ListPager uses.